requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
PyPDF2>=3.0.0
pdfplumber>=0.9.0
supabase>=2.0.0 